A snapshot of something that doesn't quite work.
This is exasperating:

I have code that works perfectly well on the sample input but doesn't quite pass muster for the file input.

I accidentally typed the right answer for the first part of the problem—386. I was off by 2: The test's output is actually 388. (I augmented the 'succeeds' test to check for this.)

I proceeded to try part 2. Once again, the sample input processes correctly, but the answer isn't correct. (And this time, it wasn't off by 2.)

I'm now going open-book to see where I've screwed up.